@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "ipyannotator"
-version = "0.4.0"
+version = "0.8.1"
 description = "The infinitely hackable annotation framework"
 authors = ["palaimon.io <oss@mail.palaimon.io>"]
 license = "Apache License 2.0"
@@ -8,23 +8,50 @@ readme = "README.md"
 repository = "https://github.com/palaimon/ipyannotator"
 
 [tool.poetry.dependencies]
-python = "^3.7"
+python = "^3.8"
 pandas = "^1.1.2"
-traitlets = '=5.0.0'
-ipycanvas = "^0.5.1"
+traitlets = "^5.1.1"
+ipycanvas = "^0.10.2"
 ipyevents = "^0.8.0"
 ipykernel = "^5.3.4"
 ipywidgets = "^7.5.1"
-jupyterlab = "^2.2.8"
-twine = "^3.2.0"
-wheel = "=0.36.2"
+scikit-image = "^0.18.3"
+matplotlib = "^3.4.3"
+pydantic = "^1.8.2"
+PyPubSub = "^4.0.3"
+pooch = "^1.5.2"
+ipython = "^8.0.1"
+nbdev = "^1.2.0"
+voila = "^0.3.1"
+tqdm = "^4.64.0"
 
 [tool.poetry.dev-dependencies]
+jupyterlab = "^3.1.12"
+# pin jupyter-client version due failing nbdev tests
+# https://stackoverflow.com/questions/66988159/nbdev-and-coroutine-object-is-not-subscriptable
+jupyter-client = "=6.1.12"
+twine = "^3.2.0"
+wheel = "=0.36.2"
+cryptography = "=3.4.8"
 nbdev = "^1.1.0"
 nbstripout = "^0.3.9"
 RISE = "^5.6"
 nbdime = "^2.0.0"
+ipytest = "^0.12.0"
+flake8 = "^4.0.1"
+nbqa = "^1.2.3"
+pylint = "^2.12.2"
+mypy = "^0.931"
+autopep8 = "^1.6.0"
+myst-nb = "^0.13.2"
+Sphinx = "^4.4.0"
+pydata-sphinx-theme = "^0.8.0"
+
 
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.autopep8]
+ignore = ["E402","F401","F403","F405","E265","W504"]
+max_line_length = 100
