@@ -4,46 +4,41 @@ __all__ = ['get_cifar10', 'get_oxford_102_flowers', 'get_cub_200_2011']
 
 # Internal Cell
 
+import pooch
 import glob
 import json
 from pathlib import Path
 import os
-import subprocess
 import tarfile
-import urllib
 import zlib
+from typing import Dict, List, Any
 
 # Internal Cell
 
-def _download_url(url, root, filename=None):
+def _download_url(url: str, root: Path, filename: str, file_hash: str = None):
     """Download a file from a url and place it in root.
     Args:
         url (str): URL to download file from
         root (str): Directory to place downloaded file in
         filename (str, optional): Name to save the file under. If None, use the basename of the URL
+        file_hash (str, optional): Hash of the required file. If None, will always download the file
     """
-    root = os.path.expanduser(root)
+    root = Path(os.path.expanduser(root))
     if not filename:
         filename = os.path.basename(url)
-    fpath = os.path.join(root, filename)
     os.makedirs(root, exist_ok=True)
 
-    if not os.path.isfile(fpath):
-        try:
-            print('Downloading ' + url + ' to ' + fpath)
-            urllib.request.urlretrieve(url, fpath)
-        except (urllib.error.URLError, IOError) as e:
-            if url[:5] == 'https':
-                url = url.replace('https:', 'http:')
-                print('Failed download. Trying https -> http instead.'
-                        ' Downloading ' + url + ' to ' + fpath)
-                urllib.request.urlretrieve(url, fpath)
-    else:
-        print(f'File {filename} already exists, skip download.')
-
+    download_setup = pooch.create(
+        path=root,
+        base_url=url,
+        registry={
+            filename: file_hash,
+        },
+    )
+    download_setup.fetch(filename)
 
 # Internal Cell
-def _extract_tar(tar_path, output_dir):
+def _extract_tar(tar_path: Path, output_dir: Path):
     try:
         print('Extracting...')
         with tarfile.open(tar_path) as tar:
@@ -52,11 +47,18 @@ def _extract_tar(tar_path, output_dir):
         print('Failed to extract!', e)
 
 # Cell
-def get_cifar10(output_dir):
+def get_cifar10(output_dir: Path):
+    """
+    Download the cifar10 dataset.
+    """
+
     output_dir = Path(output_dir)
     dataset_dir = output_dir / 'cifar10'
 
-    _download_url(url='https://s3.amazonaws.com/fast-ai-imageclas/cifar10.tgz', root=output_dir)
+    _download_url(url='https://s3.amazonaws.com/fast-ai-imageclas/',
+                  root=output_dir, filename="cifar10.tgz",
+                  file_hash=("sha256:637c5814e11aefcb6ee76d5f"
+                             "59c67ddc8de7f5b5077502a195b0833d1e3e4441"))
 
     if not dataset_dir.is_dir():
         _extract_tar(output_dir / 'cifar10.tgz', output_dir)
@@ -88,27 +90,41 @@ def get_cifar10(output_dir):
     return train_path, test_path
 
 # Cell
-def get_oxford_102_flowers(output_dir):
+def get_oxford_102_flowers(output_dir: Path):
+    """
+    Download the oxford flowers dataset.
+    """
     output_dir = Path(output_dir)
     dataset_dir = output_dir / 'oxford-102-flowers'
 
-    _download_url(url='https://s3.amazonaws.com/fast-ai-imageclas/oxford-102-flowers.tgz', root=output_dir)
+    _download_url(url='https://s3.amazonaws.com/fast-ai-imageclas/',
+                  root=output_dir, filename="oxford-102-flowers.tgz",
+                  file_hash=("sha256:680a253086535b2c800aada76a45fc1"
+                             "89d3dcae4da6da8db36ce6a95f00bf4ad"))
 
     if not dataset_dir.is_dir():
         _extract_tar(output_dir / 'oxford-102-flowers.tgz', output_dir)
     else:
-        print(f'Directory {dataset_dir} already exists, skip extraction.')
+        print((f'Directory {dataset_dir} already'
+               ' exists, skip extraction.'))
 
     print('Generating train/test data..')
     with open(dataset_dir / 'train.txt', 'r') as f:
-        annotations_train = dict(tuple(line.split()) for line in f)
+        # https://github.com/python/mypy/issues/7558
+        _annotations_train: Dict[str, Any] = dict(tuple(
+            line.split()) for line in f)  # type: ignore
 
-    annotations_train = {str(dataset_dir / k): [v+'.jpg'] for k, v in annotations_train.items()}
+    annotations_train: Dict[str, List[str]] = {
+        str(dataset_dir / k): [v + '.jpg'] for k, v in _annotations_train.items()
+    }
 
     with open(dataset_dir / 'test.txt', 'r') as f:
-        annotations_test = dict(tuple(line.split()) for line in f)
+        _annotations_test: Dict[str, Any] = dict(tuple(
+            line.split()) for line in f)  # type: ignore
 
-    annotations_test = {str(dataset_dir / k): [v+'.jpg'] for k, v in annotations_test.items()}
+    annotations_test: Dict[str, List[str]] = {
+        str(dataset_dir / k): [v + '.jpg'] for k, v in _annotations_test.items()
+    }
 
     train_path = dataset_dir / 'annotations_train.json'
     test_path = dataset_dir / 'annotations_test.json'
@@ -122,11 +138,17 @@ def get_oxford_102_flowers(output_dir):
     return train_path, test_path
 
 # Cell
-def get_cub_200_2011(output_dir):
+def get_cub_200_2011(output_dir: Path):
+    """
+    Download the CUB 200 2001 dataset.
+    """
     output_dir = Path(output_dir)
     dataset_dir = output_dir / 'CUB_200_2011'
 
-    _download_url(url='https://s3.amazonaws.com/fast-ai-imageclas/CUB_200_2011.tgz', root=output_dir)
+    _download_url(url='https://s3.amazonaws.com/fast-ai-imageclas/',
+                  root=output_dir, filename='CUB_200_2011.tgz',
+                  file_hash=("sha256:0c685df5597a8b24909f6a7c9db6"
+                             "d11e008733779a671760afef78feb49bf081"))
 
     if not dataset_dir.is_dir():
         _extract_tar(output_dir / 'CUB_200_2011.tgz', output_dir)
@@ -134,27 +156,36 @@ def get_cub_200_2011(output_dir):
         print(f'Directory {dataset_dir} already exists, skip extraction.')
 
     print('Generating train/test data..')
-    with open(dataset_dir / 'images.txt','r') as f:
-        image_id_map = dict(tuple(line.split()) for line in f)
+    with open(dataset_dir / 'images.txt', 'r') as f:
+        image_id_map: Dict[str, Any] = dict(tuple(
+            line.split()) for line in f)  # type: ignore
 
-    with open(dataset_dir / 'classes.txt','r') as f:
-        class_id_map = dict(tuple(line.split()) for line in f)
+    with open(dataset_dir / 'classes.txt', 'r') as f:
+        class_id_map: Dict[str, Any] = dict(tuple(
+            line.split()) for line in f)  # type: ignore
 
-    with open(dataset_dir / 'train_test_split.txt','r') as f:
-        splitter = dict(tuple(line.split()) for line in f)
+    with open(dataset_dir / 'train_test_split.txt', 'r') as f:
+        splitter: Dict[str, Any] = dict(tuple(
+            line.split()) for line in f)  # type: ignore
 
     # image ids for test/train
     train_k = [k for k, v in splitter.items() if v == '0']
     test_k = [k for k, v in splitter.items() if v == '1']
 
-    with open(dataset_dir / 'image_class_labels.txt','r') as f:
-        anno_ = dict(tuple(line.split()) for line in f)
+    with open(dataset_dir / 'image_class_labels.txt', 'r') as f:
+        anno_: Dict[str, Any] = dict(tuple(
+            line.split()) for line in f)  # type: ignore
 
-    annotations_train = {str(dataset_dir / 'images' / image_id_map[k]): [class_id_map[v]+'.jpg'] for k, v in anno_.items() if k in train_k}
-    annotations_test = {str(dataset_dir / 'images' / image_id_map[k]): [class_id_map[v]+'.jpg'] for k, v in anno_.items() if k in test_k}
+    annotations_train = {
+        str(dataset_dir / 'images' / image_id_map[k]):
+        [class_id_map[v] + '.jpg'] for k, v in anno_.items() if k in train_k}
 
-    train_path = dataset_dir  / 'annotations_train.json'
-    test_path = dataset_dir  / 'annotations_test.json'
+    annotations_test = {
+        str(dataset_dir / 'images' / image_id_map[k]):
+        [class_id_map[v] + '.jpg'] for k, v in anno_.items() if k in test_k}
+
+    train_path = dataset_dir / 'annotations_train.json'
+    test_path = dataset_dir / 'annotations_test.json'
 
     with open(train_path, 'w') as f:
         json.dump(annotations_train, f)
