@@ -2,60 +2,180 @@
 
 __all__ = ["index", "modules", "custom_doc_links", "git_url"]
 
-index = {"draw_bg": "01_bbox_canvas.ipynb",
+index = {"StateSettings": "00_base.ipynb",
+         "BaseState": "00_base.ipynb",
+         "AnnotatorStep": "00_base.ipynb",
+         "AppWidgetState": "00_base.ipynb",
+         "Annotator": "00a_annotator.ipynb",
+         "Settings": "00_base.ipynb",
+         "generate_subset_anno_json": "00_base.ipynb",
+         "validate_project_path": "00_base.ipynb",
+         "AnnotatorFactory": "00a_annotator.ipynb",
+         "Bboxer": "00a_annotator.ipynb",
+         "Im2Imer": "00a_annotator.ipynb",
+         "Capturer": "00a_annotator.ipynb",
+         "ImExplorer": "00a_annotator.ipynb",
+         "VideoBboxer": "00a_annotator.ipynb",
+         "Coordinate": "00b_mltypes.ipynb",
+         "BboxCoordinate": "00b_mltypes.ipynb",
+         "BboxVideoCoordinate": "00b_mltypes.ipynb",
+         "Input": "00b_mltypes.ipynb",
+         "Output": "00b_mltypes.ipynb",
+         "InputImage": "00b_mltypes.ipynb",
+         "OutputImageLabel": "00b_mltypes.ipynb",
+         "OutputLabel": "00b_mltypes.ipynb",
+         "OutputImageBbox": "00b_mltypes.ipynb",
+         "OutputVideoBbox": "00b_mltypes.ipynb",
+         "OutputGridBox": "00b_mltypes.ipynb",
+         "NoOutput": "00b_mltypes.ipynb",
+         "AnnotationStore": "00c_annotation_types.ipynb",
+         "LabelStore": "00c_annotation_types.ipynb",
+         "LabelStoreCaster": "00c_annotation_types.ipynb",
+         "is_building_docs": "00d_doc_utils.ipynb",
+         "nbglob": "00d_doc_utils.ipynb",
+         "upd_metadata": "00d_doc_utils.ipynb",
+         "hide": "00d_doc_utils.ipynb",
+         "collapse_cells": "00d_doc_utils.ipynb",
+         "draw_bg": "01_bbox_canvas.ipynb",
          "draw_bounding_box": "01_bbox_canvas.ipynb",
+         "BoundingBox": "01_bbox_canvas.ipynb",
          "get_image_size": "01_bbox_canvas.ipynb",
-         "draw_img": "01_bbox_canvas.ipynb",
+         "ImageCanvas": "01_bbox_canvas.ipynb",
+         "ImageCanvasPrototype": "01_bbox_canvas.ipynb",
+         "CanvasScaleMixin": "01_bbox_canvas.ipynb",
+         "ScaledImage": "01_bbox_canvas.ipynb",
+         "FitImage": "01_bbox_canvas.ipynb",
+         "ImageRenderer": "01_bbox_canvas.ipynb",
          "points2bbox_coords": "01_bbox_canvas.ipynb",
+         "coords_scaled": "01_bbox_canvas.ipynb",
+         "BBoxLayer": "01_bbox_canvas.ipynb",
+         "BBoxCanvasState": "01_bbox_canvas.ipynb",
+         "BBoxCanvasGUI": "01_bbox_canvas.ipynb",
+         "BBoxVideoCanvasGUI": "01_bbox_canvas.ipynb",
+         "BBoxCanvasController": "01_bbox_canvas.ipynb",
+         "BBoxVideoCanvasController": "01_bbox_canvas.ipynb",
          "BBoxCanvas": "01_bbox_canvas.ipynb",
-         "draw_bbox": "01a_datasets.ipynb",
-         "xywh_to_xyxy": "01a_datasets.ipynb",
-         "xyxy_to_xywh": "01a_datasets.ipynb",
-         "bbox_intersection": "01a_datasets.ipynb",
-         "overlap": "01a_datasets.ipynb",
-         "sample_bbox": "01a_datasets.ipynb",
-         "draw_rectangle": "01a_datasets.ipynb",
-         "draw_ellipse": "01a_datasets.ipynb",
-         "create_simple_object_detection_dataset": "01a_datasets.ipynb",
-         "create_color_classification": "01a_datasets.ipynb",
-         "create_shape_color_classification": "01a_datasets.ipynb",
-         "create_object_detection": "01a_datasets.ipynb",
+         "BBoxVideoCanvas": "01_bbox_canvas.ipynb",
+         "flatten": "01_helpers.ipynb",
+         "text_on_img": "01_helpers.ipynb",
+         "reconstruct_class_images": "01_helpers.ipynb",
+         "augment": "01_helpers.ipynb",
+         "Tutorial": "01_helpers.ipynb",
+         "show_img": "01a_datasets.ipynb",
+         "save_img": "01a_datasets.ipynb",
+         "save_img_annotations": "01a_datasets.ipynb",
+         "draw_grid": "01a_datasets.ipynb",
+         "draw_bbox": "13_datasets_legacy.ipynb",
+         "xywh_to_xyxy": "13_datasets_legacy.ipynb",
+         "xyxy_to_xywh": "13_datasets_legacy.ipynb",
+         "bbox_intersection": "13_datasets_legacy.ipynb",
+         "overlap": "13_datasets_legacy.ipynb",
+         "bb_intersection_over_union": "13_datasets_legacy.ipynb",
+         "sample_bbox": "13_datasets_legacy.ipynb",
+         "draw_rectangle": "13_datasets_legacy.ipynb",
+         "draw_ellipse": "13_datasets_legacy.ipynb",
+         "create_simple_object_detection_dataset": "13_datasets_legacy.ipynb",
+         "create_color_classification": "13_datasets_legacy.ipynb",
+         "create_shape_color_classification": "13_datasets_legacy.ipynb",
+         "create_object_detection": "13_datasets_legacy.ipynb",
          "get_cifar10": "01a_datasets_download.ipynb",
          "get_oxford_102_flowers": "01a_datasets_download.ipynb",
          "get_cub_200_2011": "01a_datasets_download.ipynb",
+         "DS": "14_datasets_factory_legacy.ipynb",
+         "get_settings": "14_datasets_factory_legacy.ipynb",
+         "TrackObject": "01b_dataset_video.ipynb",
+         "TrackObjectRectangle": "01b_dataset_video.ipynb",
+         "TrackObjectCircle": "01b_dataset_video.ipynb",
+         "Frame": "01b_dataset_video.ipynb",
+         "gen_frame": "01b_dataset_video.ipynb",
+         "gen_bg_img": "01b_dataset_video.ipynb",
+         "default_path": "01b_dataset_video.ipynb",
+         "gen_moving_object": "01b_dataset_video.ipynb",
+         "gen_render_objects": "01b_dataset_video.ipynb",
+         "occlude_lower_left_img": "01b_dataset_video.ipynb",
+         "gt_to_mot": "01b_dataset_video.ipynb",
+         "create_mot_ds": "01b_dataset_video.ipynb",
          "NaviGUI": "02_navi_widget.ipynb",
          "NaviLogic": "02_navi_widget.ipynb",
          "Navi": "02_navi_widget.ipynb",
+         "BBoxItem": "02a_right_menu_widget.ipynb",
+         "BBoxVideoItem": "02a_right_menu_widget.ipynb",
+         "BBoxList": "02a_right_menu_widget.ipynb",
+         "BBoxVideoList": "02a_right_menu_widget.ipynb",
+         "Grid": "02b_grid_menu.ipynb",
+         "GridMenu": "02b_grid_menu.ipynb",
+         "group_files_by_class": "03_storage.ipynb",
+         "construct_annotation_path": "03_storage.ipynb",
          "setup_project_paths": "03_storage.ipynb",
          "get_image_list_from_folder": "03_storage.ipynb",
+         "strip_path": "03_storage.ipynb",
+         "MapeableStorage": "03_storage.ipynb",
          "AnnotationStorage": "03_storage.ipynb",
+         "JsonLabelStorage": "03_storage.ipynb",
+         "JsonCaptureStorage": "03_storage.ipynb",
          "AnnotationStorageIterator": "03_storage.ipynb",
          "AnnotationDBStorage": "03_storage.ipynb",
+         "BBoxState": "04_bbox_annotator.ipynb",
+         "BBoxCoordinates": "04_bbox_annotator.ipynb",
          "BBoxAnnotatorGUI": "04_bbox_annotator.ipynb",
-         "BBoxAnnotatorLogic": "04_bbox_annotator.ipynb",
+         "BBoxAnnotatorController": "04_bbox_annotator.ipynb",
          "BBoxAnnotator": "04_bbox_annotator.ipynb",
+         "ImageButtonSetting": "05_image_button.ipynb",
          "ImageButton": "05_image_button.ipynb",
-         "CaptureGrid": "06_capture_annotator.ipynb",
+         "CaptureState": "06_capture_annotator.ipynb",
          "CaptureAnnotatorGUI": "06_capture_annotator.ipynb",
-         "CaptureAnnotatorLogic": "06_capture_annotator.ipynb",
+         "CaptureAnnotationStorage": "06_capture_annotator.ipynb",
+         "CaptureAnnotatorController": "06_capture_annotator.ipynb",
          "CaptureAnnotator": "06_capture_annotator.ipynb",
-         "ImCanvas": "07_im2im_annotator.ipynb",
+         "Im2ImState": "07_im2im_annotator.ipynb",
          "Im2ImAnnotatorGUI": "07_im2im_annotator.ipynb",
-         "text_on_img": "07_im2im_annotator.ipynb",
-         "flatten": "07_im2im_annotator.ipynb",
-         "reconstruct_class_images": "07_im2im_annotator.ipynb",
-         "Im2ImAnnotatorLogic": "07_im2im_annotator.ipynb",
-         "Im2ImAnnotator": "07_im2im_annotator.ipynb"}
+         "Im2ImAnnotatorController": "07_im2im_annotator.ipynb",
+         "Im2ImAnnotator": "07_im2im_annotator.ipynb",
+         "OutputWidgetHandler": "12_debug_utils.ipynb",
+         "IpyLogger": "12_debug_utils.ipynb",
+         "CoordinateInput": "15_coordinates_input.ipynb",
+         "ActionButton": "16_custom_buttons.ipynb",
+         "ExploreAnnotatorState": "17_annotator_explorer.ipynb",
+         "ExploreAnnotatorGUI": "17_annotator_explorer.ipynb",
+         "Storage": "17_annotator_explorer.ipynb",
+         "InMemoryStorage": "17_annotator_explorer.ipynb",
+         "ExploreAnnotatorController": "17_annotator_explorer.ipynb",
+         "ExploreAnnotator": "17_annotator_explorer.ipynb",
+         "TrajectoryStore": "18_bbox_trajectory.ipynb",
+         "BBoxTrajectory": "18_bbox_trajectory.ipynb",
+         "BboxVideoCoordSelected": "19_bbox_video_annotator.ipynb",
+         "BboxVideoHistory": "19_bbox_video_annotator.ipynb",
+         "BBoxVideoState": "19_bbox_video_annotator.ipynb",
+         "BBoxVideoCoordinates": "19_bbox_video_annotator.ipynb",
+         "BBoxAnnotatorVideoGUI": "19_bbox_video_annotator.ipynb",
+         "BBoxVideoAnnotatorController": "19_bbox_video_annotator.ipynb",
+         "BBoxVideoAnnotator": "19_bbox_video_annotator.ipynb"}
 
-modules = ["bbox_canvas.py",
+modules = ["base.py",
+           "annotator.py",
+           "mltypes.py",
+           "ipytyping/annotations.py",
+           "doc_utils.py",
+           "bbox_canvas.py",
+           "helpers.py",
            "datasets/generators.py",
            "datasets/download.py",
+           "datasets/factory.py",
            "navi_widget.py",
+           "right_menu_widget.py",
+           "custom_widgets/grid_menu.py",
            "storage.py",
            "bbox_annotator.py",
-           "image_button.py",
+           "custom_input/buttons.py",
            "capture_annotator.py",
-           "im2im_annotator.py"]
+           "im2im_annotator.py",
+           "debug_utils.py",
+           "datasets/generators_legacy.py",
+           "datasets/factory_legacy.py",
+           "custom_input/coordinates.py",
+           "explore_annotator.py",
+           "services/bbox_trajectory.py",
+           "bbox_video_annotator.py"]
 
 doc_url = "//ipyannotator"
 
