@@ -1,115 +1,60 @@
 # AUTOGENERATED! DO NOT EDIT! File to edit: nbs/06_capture_annotator.ipynb (unless otherwise specified).
 
-__all__ = ['CaptureGrid', 'CaptureAnnotator']
+__all__ = ['CaptureAnnotator']
 
 # Internal Cell
 
-from functools import partial
 import math
-
-from ipywidgets import (AppLayout, VBox, HBox, Button, GridBox, Layout, Checkbox, HTML, IntText, Valid, Output)
-from ipyevents import Event
-
+import warnings
+from copy import deepcopy
 from pathlib import Path
+from typing import Dict, Optional, Callable, List
 
+from IPython.display import display
+from ipywidgets import (AppLayout, HBox, Button, HTML, VBox,
+                        Layout, Checkbox, Output)
+
+from .custom_widgets.grid_menu import GridMenu, Grid
+from .base import BaseState, AppWidgetState, Annotator
 from .navi_widget import Navi
-from .storage import setup_project_paths, get_image_list_from_folder, AnnotationStorage
+from .ipytyping.annotations import LabelStore, _label_store_to_image_button
+from .storage import JsonCaptureStorage
 
-from traitlets import Dict, Int, HasTraits, observe, dlink, link, List, Unicode, Bool
-
-from .image_button import ImageButton
-
-# Cell
-
-class CaptureGrid(GridBox, HasTraits):
-    """
-    Represents grid of `ImageButtons` with state.
-
-    """
-    debug_output = Output(layout={'border': '1px solid black'})
-    current_state = Dict()
-
-    def __init__(self, grid_item=ImageButton, image_width=150, image_height=150,
-                 n_rows=3, n_cols=3, display_label=False):
-
-
-        self.image_width = image_width
-        self.image_height = image_height
-        self.n_rows = n_rows
-        self.n_cols = n_cols
-        self._screen_im_number = IntText(value=n_rows * n_cols,
-                                         description='screen_image_number',
-                                         disabled=False)
-
-        self._labels = [grid_item(display_label=display_label,
-                                  image_width='%dpx' % self.image_width,
-                                  image_height='%dpx' % self.image_height) for _ in range(self._screen_im_number.value)]
-
-        self.callback = None
-
-        self.observe(self.on_state_change, 'current_state')
-
-
-        gap = 40 if display_label else 15
-
-        centered_settings = {
-            'grid_template_columns': " ".join(["%dpx" % (self.image_width + gap) for i
-                                               in range(self.n_cols)]),
-            'grid_template_rows': " ".join(["%dpx" % (self.image_height + gap) for i
-                                            in range(self.n_rows)]),
-            'justify_content': 'center',
-            'align_content': 'space-around'
-        }
-
-
-        super().__init__(children=self._labels, layout=Layout(**centered_settings))
-
-
-    @debug_output.capture(clear_output=True)
-    def on_state_change(self, change=None):
-        print('on_state_change', change['new'])
-        new_state = change['new']
-        updated = 0
-        iter_state = iter(new_state)
-
-        for label in self._labels:
-            p = next(iter_state, None)
-            if p:
-                label.image_path = str(p)
-                label.label_value = Path(p).stem
-                label.active = new_state[p].get('answer', False)
-            else:
-                label.clear()
-
-        if self.callback:
-            self.register_on_click()
-
-
-    def on_click(self, cb):
-        self.callback = cb
-        self.register_on_click()
-
-    @debug_output.capture(clear_output=True)
-    def register_on_click(self):
-        print('register_on_click')
-        for l in self._labels:
-            l.reset_callbacks()
-            l.on_click(partial(self.callback, name=l.name))
+# Internal Cell
+class CaptureState(BaseState):
+    annotations: LabelStore = LabelStore()
+    grid: Grid
+    question_value: str = ''
+    all_none: bool = False
 
 # Internal Cell
 
 class CaptureAnnotatorGUI(AppLayout):
-    def __init__(self, image_width=150, image_height=150,
-                 n_rows=3, n_cols=3):
+    """
+    save_btn_clicked: callable
+        activated when the user clicked on the save button
+    grid_box_clicked: callable
+        activated when the user clicked on the grid box
+    on_navi_clicked: callable
+        activated when the user navigates through the annotator
+    """
 
-        self._screen_im_number = IntText(value=n_rows * n_cols,
-                                    description='screen_image_number',
-                                    disabled=False)
+    debug_output = Output(layout={'border': '1px solid black'})
 
-        self.image_width = image_width
-        self.image_height = image_height
-        self.n_rows = n_rows
-        self.n_cols = n_cols
+    def __init__(
+        self,
+        app_state: AppWidgetState,
+        capture_state: CaptureState,
+        save_btn_clicked: Callable = None,
+        grid_box_clicked: Callable = None,
+        on_navi_clicked: Callable = None,
+        select_none_changed: Callable = None
+    ):
+        self._app_state = app_state
+        self._capture_state = capture_state
+        self._save_btn_clicked = save_btn_clicked
+        self._grid_box_clicked = grid_box_clicked
+        self._select_none_changed = select_none_changed
 
         self._navi = Navi()
 
@@ -120,115 +65,258 @@ class CaptureAnnotatorGUI(AppLayout):
                                        indent=False,
                                        layout=Layout(width='100px'))
 
-        self._controls_box = HBox([self._navi, self._save_btn, self._none_checkbox],
-                                 layout=Layout(display='flex', justify_content='center', flex_flow='wrap', align_items='center'))
+        self._controls_box = HBox(
+            [
+                self._navi,
+                self._save_btn,
+                self._none_checkbox,
+            ],
+            layout=Layout(
+                display='flex',
+                justify_content='center',
+                flex_flow='wrap',
+                align_items='center'
+            )
+        )
 
-
-        self._grid_box = CaptureGrid(image_width=image_width, image_height=image_height,  n_rows=n_rows, n_cols=n_cols, display_label=False)
-
+        self._grid_box = GridMenu(capture_state.grid)
 
         self._grid_label = HTML()
-        self._labels_box = VBox(children = [self._grid_label, self._grid_box],
-                                layout=Layout(display='flex', justify_content='center', flex_wrap='wrap', align_items='center'))
+        self._labels_box = VBox(
+            children=[
+                self._grid_label,
+                self._grid_box
+            ],
+            layout=Layout(
+                display='flex',
+                justify_content='center',
+                flex_wrap='wrap',
+                align_items='center'
+            )
+        )
 
+        self.on_navi_clicked = on_navi_clicked
+        self._navi.on_navi_clicked = self._on_navi_clicked
+        self._save_btn.on_click(self._btn_clicked)
+        self._grid_box.on_click(self.on_grid_clicked)
+        self._none_checkbox.observe(self._none_checkbox_changed, 'value')
 
-        super().__init__(header=None,
-                 left_sidebar=None,
-                 center=self._labels_box,
-                 right_sidebar=None,
-                 footer=self._controls_box,
-                 pane_widths=(2, 8, 0),
-                 pane_heights=(1, 4, 1))
+        if self._capture_state.question_value:
+            self._set_label(self._capture_state.question_value)
 
+        if self._app_state.max_im_number:
+            self._set_navi_max_im_number(self._app_state.max_im_number)
+
+        if self._capture_state.annotations:
+            self._load_menu(self._capture_state.annotations)
+
+        self._capture_state.subscribe(self._set_none_checkbox, 'all_none')
+        self._capture_state.subscribe(self._set_label, 'question_value')
+        self._app_state.subscribe(self._set_navi_max_im_number, 'max_im_number')
+        self._capture_state.subscribe(self._load_menu, 'annotations')
+
+        super().__init__(
+            header=None,
+            left_sidebar=None,
+            center=self._labels_box,
+            right_sidebar=None,
+            footer=self._controls_box,
+            pane_widths=(2, 8, 0),
+            pane_heights=(1, 4, 1))
+
+    def _on_navi_clicked(self, index: int):
+        if self.on_navi_clicked:
+            self.on_navi_clicked(index)
+
+        self._grid_box.load(
+            _label_store_to_image_button(self._capture_state.annotations)
+        )
+
+    @debug_output.capture(clear_output=True)
+    def _load_menu(self, annotations: LabelStore):
+        self._grid_box.load(
+            _label_store_to_image_button(annotations)
+        )
+
+    def _set_none_checkbox(self, all_none: bool):
+        self._none_checkbox.value = all_none
+
+    def _set_navi_max_im_number(self, max_im_number: int):
+        self._navi.max_im_num = max_im_number
+
+    def _set_label(self, question_value: str):
+        self._grid_label.value = question_value
+
+    def _btn_clicked(self, *args):
+        if self._save_btn_clicked:
+            self._save_btn_clicked(self._app_state.index)
+        else:
+            warnings.warn("Save button click didn't triggered any event.")
+
+    def _none_checkbox_changed(self, change: dict):
+        self._capture_state.set_quietly('all_none', change['new'])
+        if self._select_none_changed:
+            self._select_none_changed(change)
+
+    def on_grid_clicked(self, event, name=None):
+        if self._grid_box_clicked:
+            self._grid_box_clicked(event, name)
+        else:
+            warnings.warn("Grid box click didn't triggered any event.")
 
 # Internal Cell
 
-class CaptureAnnotatorLogic(HasTraits):
-    debug_output = Output(layout={'border': '1px solid black'})
-    index = Int(0) # state index
-    disp_number = Int() # number of images on screen
-    num_screens = Int() # number of screens
-    current_state = Dict()
-    question_value = Unicode()
-    all_none = Bool()
+class CaptureAnnotationStorage:
+    def __init__(
+        self,
+        input_item_path: Path,
+        annotation_file_path: str
+    ):
+        self.input_item_path = input_item_path
+        self.annotation_file_path = annotation_file_path
 
+        self.annotations = JsonCaptureStorage(
+            self.input_item_path,
+            annotation_file_path=self.annotation_file_path
+        )
 
-    def __init__(self, project_path, question=None, image_dir='pics', filter_files=None, results_dir=None):
-        self.project_path = Path(project_path)
-        self.image_dir, self.annotation_file_path = setup_project_paths(self.project_path, image_dir=image_dir, results_dir=results_dir)
+    def __setitem__(self, key, value):
+        self.annotations[key] = value
 
-        self.image_paths = sorted(get_image_list_from_folder(self.image_dir)) #todo: use sorted in storage?
+    def _save(self):
+        self.annotations.save()
 
-        if filter_files:
-            self.image_paths = [p for p in self.image_paths if str(p) in filter_files]
-        if not self.image_paths:
-            raise UserWarning("No image files to display! Check image_dir of filter.")
-        self.current_im_num = len(self.image_paths)
-        self.annotations = AnnotationStorage(self.image_paths)
-        if question:
-            self.question_value = f'<center><p style="font-size:20px;">{question}</p></center>'
-        self._update_state()
+    def get_im_names(self, filter_files) -> List[str]:
+        return self.annotations.get_im_names(filter_files)
 
+    def get(self, path: str) -> Optional[Dict]:
+        return self.annotations.get(path)
 
-    @observe('disp_number')
-    def _update_state(self, change=None): # from annotations
-        state_images = self._get_state_names(self.index)
-        current_state = {}
-        for im_path in state_images:
-            current_state[str(im_path)] = self.annotations.get(str(im_path)) or {}
-        self.all_none = all(value == {'answer': False} for value in current_state.values())
-        self.current_state = current_state
-
-
-    def _update_annotations(self, index): # from screen
-        for p, anno in self.current_state.items():
+    def update_annotations(self, annotations: dict):
+        for p, anno in annotations.items():
             self.annotations[str(p)] = anno
+        self._save()
 
-    def _save_annotations(self, *args, **kwargs): # to disk
-        index = kwargs.pop('old_index', self.index)
-        self._update_annotations(index)
-        self.annotations.save(self.annotation_file_path)
+    def to_dict(self, only_annotated: bool = True) -> dict:
+        return self.annotations.to_dict(only_annotated)
 
-    def _get_state_names(self, index):
-        start = index * self.disp_number
-        end = start + self.disp_number
-        im_names = self.image_paths[start:end]
+# Internal Cell
+
+class CaptureAnnotatorController:
+    debug_output = Output(layout={'border': '1px solid black'})
+
+    def __init__(
+        self,
+        app_state: AppWidgetState,
+        capture_state: CaptureState,
+        storage: CaptureAnnotationStorage,
+        input_item=None,
+        output_item=None,
+        filter_files=None,
+        question=None,
+        *args,
+        **kwargs
+    ):
+        self._app_state = app_state
+        self._capture_state = capture_state
+        self._storage = storage
+        self.input_item = input_item
+        self.output_item = output_item
+        self._last_index = 0
+
+        self.images = self._storage.get_im_names(filter_files)
+        self.current_im_number = len(self.images)
+
+        if question:
+            self._capture_state.question_value = ('<center><p style="font-size:20px;"'
+                                                  f'>{question}</p></center>')
+
+        self.update_state()
+        self._calc_screens_num()
+
+    def update_state(self):
+        state_images = self._get_state_names(self._app_state.index)
+        tmp_annotations = deepcopy(self._capture_state.annotations)
+        current_state = {}
+
+        for im_path in state_images:
+            current_state[str(im_path)] = self._storage.get(str(im_path)) or {}
+
+        self._update_all_none_state(current_state)
+
+        # error: Incompatible types in assignment (expression has type
+        # "Dict[str, Dict[Any, Any]]", variable has type
+        # "Dict[str, Optional[Dict[str, bool]]]")
+        tmp_annotations.clear()
+        tmp_annotations.update(current_state)
+        self._capture_state.annotations = tmp_annotations  # type: ignore
+
+    def _update_all_none_state(self, state_images: dict):
+        self._capture_state.all_none = all(
+            value == {'answer': False} for value in state_images.values()
+        )
+
+    def save_annotations(self, index: int):  # to disk
+        state_images = dict(self._capture_state.annotations)
+
+        self._storage.update_annotations(state_images)
+
+    def _get_state_names(self, index: int) -> List[str]:
+        start = index * self._capture_state.grid.disp_number
+        end = start + self._capture_state.grid.disp_number
+        im_names = self.images[start:end]
         return im_names
 
-
-    @observe('index')
-    def _idx_changed(self, change):
+    def idx_changed(self, index: int):
         ''' On index change save old state
             and update current state for visualisation
         '''
-        self._save_annotations(old_index = change['old'])
-        # update new screen
-        self._update_state()
+        self._app_state.set_quietly('index', index)
+        self.save_annotations(self._last_index)
+        self.update_state()
+        self._last_index = index
 
-    @observe('disp_number')
-    def _calc_screens_num(self, change=None):
-        self.num_screens = math.ceil(self.current_im_num / self.disp_number)
-
+    def _calc_screens_num(self):
+        self._app_state.max_im_number = math.ceil(
+            self.current_im_number / self._capture_state.grid.disp_number
+        )
 
     @debug_output.capture(clear_output=False)
-    def _handle_grid_click(self, event, name=None):
-        p = Path(self.image_dir, name)
-        current_state = self.current_state.copy()
+    def handle_grid_click(self, event: dict, name=None):
+        p = self._storage.input_item_path / name
+        current_state = deepcopy(self._capture_state.annotations)
         if not p.is_dir():
-            current_state[str(p)] = {'answer': not self.current_state[str(p)].get('answer', False)}
-            if self.all_none:
-                self.all_none = False
+            state_answer = self._capture_state.annotations[
+                str(p)].get('answer', False)
+            current_state[str(p)] = {'answer': not state_answer}
+
+            for k, v in current_state.items():
+                if v == {}:
+                    current_state[k] = {'answer': False}
+
+            if self._capture_state.all_none:
+                self._capture_state.all_none = False
+            else:
+                self._update_all_none_state(dict(current_state))
         else:
             return
-        self.current_state = current_state
 
-    def _select_none(self, change=None):
-        self.current_state = {p: {'answer': False} for p in self.current_state}
+        self._capture_state.annotations = current_state
+
+    def select_none(self, change: dict):
+        if self._capture_state.all_none:
+            tmp_annotations = deepcopy(self._capture_state.annotations)
+            tmp_annotations.clear()
+            tmp_annotations.update(
+                {p: {
+                    'answer': False} for p in self._capture_state.annotations}
+            )
+            self._capture_state.annotations = tmp_annotations
 
 # Cell
 
-
-class CaptureAnnotator(CaptureAnnotatorGUI):
+class CaptureAnnotator(Annotator):
+    debug_output = Output(layout={'border': '1px solid black'})
     """
     Represents capture annotator.
 
@@ -237,39 +325,84 @@ class CaptureAnnotator(CaptureAnnotatorGUI):
     export final annotations in json format
 
     """
+#     @debug_output.capture(clear_output=True)
 
-    def __init__(self, project_path, image_dir='pics', image_width=150, image_height=150,
-                 n_rows=3, n_cols=3, question=None, filter_files=None, results_dir=None):
+    def __init__(
+        self,
+        project_path: Path,
+        input_item,
+        output_item,
+        annotation_file_path,
+        n_rows=3,
+        n_cols=3,
+        disp_number=9,
+        question=None,
+        filter_files=None
+    ):
 
-        super().__init__(image_width, image_height, n_rows, n_cols)
+        assert input_item, "WARNING: Provide valid Input"
+        assert output_item, "WARNING: Provide valid Output"
 
-        self._model = CaptureAnnotatorLogic(project_path, question, image_dir,
-                                           filter_files=filter_files, results_dir=results_dir)
+        self._project_path = project_path
+        self._input_item = input_item
+        self._output_item = output_item
+        self._annotation_file_path = annotation_file_path
+        self._n_rows = n_rows
+        self._n_cols = n_cols
+        self._question = question
+        self._filter_files = filter_files
 
-        self._save_btn.on_click(self._model._save_annotations)
+        app_state = AppWidgetState(
+            uuid=str(id(self)),
+            **{'size': (input_item.width, input_item.height)}
+        )
 
-        self._grid_box.on_click(self._model._handle_grid_click)
+        super().__init__(app_state)
 
-        link((self._model, 'all_none'), (self._none_checkbox, 'value'))
+        grid = Grid(
+            width=input_item.width,
+            height=input_item.height,
+            n_rows=n_rows,
+            n_cols=n_cols,
+            display_label=False,
+            disp_number=disp_number
+        )
 
-        self._none_checkbox.observe(self._model._select_none, 'value')
+        self.capture_state = CaptureState(
+            uuid=str(id(self)),
+            annotations=LabelStore(),
+            grid=grid
+        )
 
-        # get correct screen image number from gui settings
-        dlink((self._screen_im_number, 'value'), (self._model, 'disp_number'))
+        self.storage = CaptureAnnotationStorage(
+            input_item_path=project_path / input_item.dir,
+            annotation_file_path=annotation_file_path
+        )
 
-        # set correct number of screens slider value based on image number
-        dlink((self._model, 'num_screens'), (self._navi.model, 'max_im_number'))
+        self.controller = CaptureAnnotatorController(
+            app_state=self.app_state,
+            storage=self.storage,
+            capture_state=self.capture_state,
+            input_item=input_item,
+            output_item=output_item,
+            question=question,
+            n_rows=n_rows,
+            n_cols=n_cols,
+            filter_files=filter_files
+        )
 
-        # link current image index from controls to annotator model
-        link((self._navi.model, 'index'), (self._model, 'index'))
+        self.view = CaptureAnnotatorGUI(
+            capture_state=self.capture_state,
+            app_state=self.app_state,
+            save_btn_clicked=self.controller.save_annotations,
+            grid_box_clicked=self.controller.handle_grid_click,
+            on_navi_clicked=self.controller.idx_changed,
+            select_none_changed=self.controller.select_none
+        )
 
-        # link annotation question
-        link((self._model, 'question_value'), (self._grid_label, 'value'))
+    def __repr__(self):
+        display(self.view)
+        return ""
 
-        # link state of model and grid box visualizer
-        link((self._model, 'current_state'), (self._grid_box, 'current_state'))
-
-
-
-    def to_dict(self, only_annotated=True):
-        return self._model.annotations.to_dict(only_annotated)
+    def to_dict(self, only_annotated=True) -> dict:
+        return self.storage.to_dict(only_annotated)
