@@ -1,34 +1,69 @@
 # AUTOGENERATED! DO NOT EDIT! File to edit: nbs/03_storage.ipynb (unless otherwise specified).
 
-__all__ = ['AnnotationStorage']
+__all__ = ['MapeableStorage', 'AnnotationStorage', 'AnnotationStorageIterator', 'AnnotationDBStorage']
 
-# Cell
-
-import os
-import json
-import sqlite3
+# Internal Cell
+import warnings
 import copy
-from pathlib import Path
+import json
+import os
+from typing import List, Union, Iterable
+from collections import defaultdict
 from collections.abc import MutableMapping
+from pathlib import Path
+
+# Internal Cell
+def group_files_by_class(annotations_dict):
+    grouped = defaultdict(list)
+    for file, labels in annotations_dict.items():
+        for class_ in labels:
+            grouped[class_].append(file)
+    return grouped
 
 # Internal Cell
 
-def setup_project_paths(project_path:Path, file_name=None, image_dir='pics', label_dir=None, results_dir=None):
-    assert project_path.exists(), "WARNING: Project path should point to " \
-                                       "existing directory"
-    assert project_path.is_dir(), "WARNING: Project path should point to " \
-                                       "existing directory"
-    im_dir = Path(project_path, image_dir)
+def construct_annotation_path(project_path=None, file_name=None, results_dir=None):
+    if file_name is not None:
+        annotation_file_path = Path(file_name)
+        results_dir = annotation_file_path.parent
+    elif project_path is not None:
+        results_dir = Path(
+            project_path, 'results') if results_dir is None else Path(project_path, results_dir)
+
+        annotation_file_path = Path(results_dir, 'annotations.json')
+        if annotation_file_path.is_file():
+            warnings.warn(f"Error: Annotations file already exists in {results_dir}!"
+                          "\n If you want to create annotations from scratch"
+                          " - use empty dir!")
+    else:
+        raise ValueError("You must define `project_path` or `file_name`!")
+
+    results_dir.mkdir(parents=True, exist_ok=True)
+    return annotation_file_path
+
+# Internal Cell
+from .base import validate_project_path
+
+
+def setup_project_paths(project_path, file_name=None, image_dir='pics',
+                        label_dir=None, results_dir=None):
+    project_path = validate_project_path(project_path)
+
+    im_dir = project_path / image_dir
 
     if file_name is not None:
         annotation_file_path = Path(file_name)
         results_dir = annotation_file_path.parent
         print(f"WARNING: `results_dir` is deduced from `file_name` path: {results_dir}")
     else:
-        results_dir = Path(project_path, 'results') if results_dir is None else Path(project_path, results_dir)
+        results_dir = Path(
+            project_path, 'results') if results_dir is None else Path(project_path, results_dir)
+
         annotation_file_path = Path(results_dir, 'annotations.json')
         if annotation_file_path.is_file():
-            print(f"WARNING: Annotations file already exists in {results_dir}!\n         If you want to create annotations from scratch - use empty dir isntead.")
+            print(f"WARNING: Annotations file already exists in {results_dir}"
+                  "!\n         If you want to create annotations from scratch"
+                  " - use empty dir isntead.")
 
     results_dir.mkdir(parents=True, exist_ok=True)
 
@@ -42,7 +77,8 @@ def setup_project_paths(project_path:Path, file_name=None, image_dir='pics', lab
 # Internal Cell
 import glob
 
-def get_image_list_from_folder(image_dir, strip_path=False):
+
+def get_image_list_from_folder(image_dir) -> Iterable[Path]:
     ''' Scans <image_dir> to construct list of existing images as <Path> objects
     '''
     # if no files in `image_dir` assume all images are under `class_name` directories
@@ -50,29 +86,19 @@ def get_image_list_from_folder(image_dir, strip_path=False):
         path_list = [Path(p) for p in glob.glob(f'{image_dir}/*/*')]
     else:
         path_list = [Path(image_dir, f) for f in os.listdir(image_dir) if
-                 os.path.isfile(os.path.join(image_dir, f))]
+                     os.path.isfile(os.path.join(image_dir, f))]
 
-    if strip_path:
-        path_list = [p.name for p in path_list]
     return path_list
+
+
+def strip_path(paths: Iterable[Path]) -> Iterable[str]:
+    return [p.name for p in paths]
 
 # Cell
 
-class AnnotationStorage(MutableMapping):
-    """
-    Represents generic storage for annotations.
-
-    `key` is object_id / file_name and `value` - json blob containing annotation.
-
-    im_paths - list of existing images as <Path> objects
-
-    dir_in_label - use parent folder name in label (`class/file.jpg`)
-
-    """
-    def __init__(self, im_paths, dir_in_label=False):
+class MapeableStorage(MutableMapping):
+    def __init__(self):
         self.mapping = {}
-#         self.update({os.path.join(*p.parts[-2:]) if dir_in_label else p.name: None for p in im_paths})
-        self.update({str(p): None for p in im_paths})
 
     def __getitem__(self, key):
         return self.mapping[key]
@@ -86,24 +112,151 @@ class AnnotationStorage(MutableMapping):
 
     def __iter__(self):
         return iter(self.mapping)
+
     def __len__(self):
         return len(self.mapping)
-    def __repr__(self):
-        return f"{type(self).__name__}({self.mapping})"
-
-    def save(self, file_name):
-        with open(file_name, 'w', encoding='utf-8') as f:
-            json.dump(self.mapping, f, ensure_ascii=False, sort_keys = True, indent=4)
-
-    def load(self, file_name):
-        with open(file_name) as data_file:
-            self.mapping = json.load(data_file)
 
     def to_dict(self, only_annotated=True):
         if only_annotated:
             return {k: copy.deepcopy(v) for k, v in self.mapping.items() if v}
         else:
             return copy.deepcopy(self.mapping)
+
+# Cell
+
+class AnnotationStorage(MapeableStorage):
+    """
+    Represents generic storage for annotations.
+
+    `key` is object_id / file_name and `value` - json blob containing annotation.
+
+    im_paths - list of existing images as <Path> objects
+
+    """
+
+    def __init__(self, im_paths):
+        super().__init__()
+        self.update({str(p): None for p in im_paths})
+
+    def __repr__(self):
+        return f"{type(self).__name__}({self.mapping})"
+
+    def save(self, file_name):
+        with open(file_name, 'w', encoding='utf-8') as f:
+            json.dump(self.mapping, f, ensure_ascii=False, sort_keys=True, indent=4)
+
+    def load(self, file_name):
+        with open(file_name) as data_file:
+            self.mapping = json.load(data_file)
+
+# Internal Cell
+from .helpers import flatten, reconstruct_class_images
+
+
+class JsonLabelStorage(AnnotationStorage):
+    def __init__(self, im_dir: Path, label_dir: Union[Iterable[str], Path], annotation_file_path):
+        self.annotation_file_path = annotation_file_path
+        self.label_dir = label_dir
+
+        self.has_annotation_file = True if (annotation_file_path is not None and
+                                            annotation_file_path.is_file()) else False
+
+        self.images = get_image_list_from_folder(im_dir)
+
+        if isinstance(label_dir, Path):
+            # artificialy generate labels if no class images given (TODO: temorary workaround)
+            if 'class_autogenerated_' in str(label_dir):
+                label_dir.mkdir(parents=True, exist_ok=True)
+
+                if self.has_annotation_file:
+                    reconstruct_class_images(label_dir, annotation_file_path, lbl_w=50, lbl_h=50)
+                else:
+                    warnings.warn("Annotation file should be provided"
+                                  " to generate labels automatically!")
+
+            self.labels = strip_path(get_image_list_from_folder(label_dir))
+        elif isinstance(label_dir, Iterable):
+            self.labels = label_dir
+        else:
+            raise ValueError("label_dir should have str or Path type")
+
+        if self.has_annotation_file:  # init from json
+            self.load()
+        else:  # init storage from folder
+            super().__init__(self.images)
+            self.save()
+
+    def get_im_names(self, filter_files=None):
+        keys = self.keys()
+        images = sorted([k for k in self.images if str(k) in keys])
+
+        if not images:
+            raise UserWarning("!! No Images to dipslay !!")
+
+        if filter_files is not None:
+            images = [p for p in images if str(p) in filter_files]
+
+        if not images:
+            raise UserWarning("!! No image files to display. Check filter !!")
+        return images
+
+    def get_labels(self) -> List[Union[Path, str]]:
+        if not self.labels:
+            warnings.warn("!! No labels to display !!")
+            return []
+
+        if self.has_annotation_file and isinstance(self.label_dir, Path):
+            values = set(flatten(self.values()))
+            return sorted([v for v in self.labels if str(v) in values])
+
+        # create mod -> display all labels from folder, not json
+        return sorted(self.labels)
+
+    def save(self):
+        super().save(self.annotation_file_path)
+
+    def load(self):
+        super().load(self.annotation_file_path)
+
+# Internal Cell
+
+class JsonCaptureStorage(AnnotationStorage):
+    def __init__(self, im_dir: Path, annotation_file_path):
+        self.im_dir = im_dir
+        self.annotation_file_path = annotation_file_path
+
+        self.has_annotation_file = True if (annotation_file_path is not None and
+                                            annotation_file_path.is_file()) else False
+        self.images = sorted(get_image_list_from_folder(im_dir))
+
+        if self.has_annotation_file:  # init from json
+            self.load()
+        else:  # init storage from folder
+            super().__init__(self.images)
+            self.save()
+
+    def get_im_names(self, filter_files=None):
+        images = sorted(k for k in self.images if str(k) in self.keys())
+
+        if not images:
+            raise UserWarning("!! No Images to dipslay !!")
+
+        if filter_files is not None:
+            images = [p for p in images if str(p) in filter_files]
+
+        if not images:
+            raise UserWarning("!! No image files to display. Check filter !!")
+        return images
+
+    def save(self):
+        super().save(self.annotation_file_path)
+
+    def load(self):
+        super().load(self.annotation_file_path)
+
+# Cell
+
+import sqlite3
 
 # Internal Cell
 def _list_tables(conn):
@@ -123,12 +276,12 @@ def _list_tables(conn):
 def _create_tables(conn):
     c = conn.cursor()
     query = """
-    CREATE TABLE IF NOT EXISTS data (objectID TEXT,
-                                     timestamp DATETIME DEFAULT(STRFTIME('%Y-%m-%d %H:%M:%f', 'NOW')),
-                                     data JSON,
-                                     author TEXT,
-                                     PRIMARY KEY (objectId, timestamp)
-                                     );
+CREATE TABLE IF NOT EXISTS data (objectID TEXT,
+                                 timestamp DATETIME DEFAULT(STRFTIME('%Y-%m-%d %H:%M:%f', 'NOW')),
+                                 data JSON,
+                                 author TEXT,
+                                 PRIMARY KEY (objectId, timestamp)
+                                );
     """
     c.execute(query)
     query = """
@@ -156,7 +309,7 @@ def _list_table(conn, table_name='data', latest=True):
     c = conn.cursor()
     return c.execute(query).fetchall()
 
-# Internal Cell
+# Cell
 def _get_order_id(conn, object_id, table_name='objects'):
     query = """
     SELECT orderID from {}
@@ -167,7 +320,7 @@ def _get_order_id(conn, object_id, table_name='objects'):
     if res is not None:
         return res[0]
 
-# Internal Cell
+# Cell
 def _create_order_id(conn, object_id, table_name='objects'):
     order_id = _get_order_id(conn, object_id, table_name=table_name)
     if order_id:
@@ -176,10 +329,10 @@ def _create_order_id(conn, object_id, table_name='objects'):
     INSERT INTO {}('objectID') VALUES('{}')
     """.format(table_name, object_id)
     c = conn.cursor()
-    res = c.execute(query)
+    c.execute(query)
     return _get_order_id(conn, object_id, table_name=table_name)
 
-# Internal Cell
+# Cell
 def _get(conn, object_id, table_name='data'):
     query = """
     SELECT data FROM {}
@@ -193,7 +346,7 @@ def _get(conn, object_id, table_name='data'):
     if res is not None:
         return json.loads(res[0])
 
-# Internal Cell
+# Cell
 def _get_object_id_at_pos(conn, pos, table_name='objects'):
     query = """
     SELECT objectID FROM {}
@@ -205,7 +358,7 @@ def _get_object_id_at_pos(conn, pos, table_name='objects'):
     if res is not None:
         return res[0]
 
-# Internal Cell
+# Cell
 def _insert(conn, object_id, data: dict, table_name='data', author='author'):
     # insert if values have been changed
 
@@ -217,10 +370,10 @@ def _insert(conn, object_id, data: dict, table_name='data', author='author'):
         return
     c = conn.cursor()
     c.execute("insert into {}('objectID', 'author', 'data') values (?, ?, ?)".format(table_name),
-                              [object_id, author, json.dumps(data)])
+              [object_id, author, json.dumps(data)])
     conn.commit()
 
-# Internal Cell
+# Cell
 def _to_dict(conn, table_name='data'):
     query = """
     SELECT objectID, data from {}
@@ -231,7 +384,7 @@ def _to_dict(conn, table_name='data'):
     c = conn.cursor()
     return {key: json.loads(value) for key, value in c.execute(query).fetchall()}
 
-# Internal Cell
+# Cell
 def _row_count(conn, table_name='data'):
     query = """
     SELECT COUNT(DISTINCT objectID) FROM {}
@@ -240,7 +393,7 @@ def _row_count(conn, table_name='data'):
     res = c.execute(query).fetchone()
     return res[0]
 
-# Internal Cell
+# Cell
 def _delete_last(conn, object_id, table_name='data'):
     query = """
     DELETE FROM {}
@@ -249,20 +402,20 @@ def _delete_last(conn, object_id, table_name='data'):
     LIMIT 1
     """.format(table_name, object_id)
     c = conn.cursor()
-    res = c.execute(query)
+    c.execute(query)
     conn.commit()
 
-# Internal Cell
+# Cell
 def _delete_all(conn, object_id, table_name='data'):
     query = """
     DELETE FROM {}
     WHERE objectId = '{}'
     """.format(table_name, object_id)
     c = conn.cursor()
-    res = c.execute(query)
+    c.execute(query)
     conn.commit()
 
-# Internal Cell
+# Cell
 
 class AnnotationStorageIterator:
     def __init__(self, annotator_storage):
@@ -286,7 +439,7 @@ class AnnotationStorageIterator:
             raise StopIteration
         return self.annotator_storage.at(self.index)
 
-# Internal Cell
+# Cell
 
 class AnnotationDBStorage(MutableMapping):
     def __init__(self, conn_string, im_paths=None):
